Moved `PlasmaProfile` into a dedicated module
