+++ pedinf/analysis/__init__.py
@@ -2,7 +2,7 @@ from pedinf.analysis.separatrix import separatrix_given_temperature
 from pedinf.analysis.separatrix import separatrix_given_scaling
 from pedinf.analysis.pressure import pressure_parameters, pressure_profile_and_gradient
 from pedinf.analysis.inversions import profiles_by_temperature
-from pedinf.analysis.utils import PlasmaProfile
+from pedinf.analysis.profile import PlasmaProfile
 
 __all__ = [
     "separatrix_given_temperature",
